modified dto files structure
