+++ src/main/java/com/example/intergalactic_marketplace/dto/Customer/CustomerEntry.java
@@ -1,4 +1,4 @@
-package com.example.intergalactic_marketplace.dto;
+package com.example.intergalactic_marketplace.dto.Customer;
 
 import lombok.Builder;
 import lombok.Value;
@@ -8,8 +8,8 @@ import lombok.extern.jackson.Jacksonized;
 @Builder
 @Jacksonized
 public class CustomerEntry {
-    Long id;
-    String name;
-    String address;
-    String email;
+  Long id;
+  String name;
+  String address;
+  String email;
 }
