--- src/main/java/com/example/intergalactic_marketplace/dto/OrderDto.java
@@ -1,15 +1,0 @@
-package com.example.intergalactic_marketplace.dto;
-
-import java.util.UUID;
-
-import lombok.Builder;
-import lombok.Value;
-import lombok.extern.jackson.Jacksonized;
-
-@Value
-@Builder
-@Jacksonized
-public class OrderDto {
-    UUID id;
-    ProductListDto products;
-}
