+++ src/main/java/com/example/intergalactic_marketplace/dto/Product/ProductListDto.java
@@ -1,7 +1,6 @@
-package com.example.intergalactic_marketplace.dto;
+package com.example.intergalactic_marketplace.dto.Product;
 
 import java.util.List;
-
 import lombok.Builder;
 import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
@@ -10,5 +9,5 @@ import lombok.extern.jackson.Jacksonized;
 @Builder
 @Jacksonized
 public class ProductListDto {
-    List<ProductEntry> products;
+  List<ProductEntry> products;
 }
