+++ src/main/java/com/example/intergalactic_marketplace/service/mapper/OrderMapper.java
@@ -4,7 +4,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 
 import com.example.intergalactic_marketplace.domain.Order;
-import com.example.intergalactic_marketplace.dto.OrderDto;
+import com.example.intergalactic_marketplace.dto.Order.OrderDto;
 
 @Mapper(componentModel = "spring", uses = ProductMapper.class)
 public interface OrderMapper {
